+++ dockerfiles/Dockerfile.ci.os
@@ -12,6 +12,7 @@ RUN apt-get update && \
 COPY . /app/
 
 RUN pip3 install --no-cache-dir -e ./stac_fastapi/core && \
+    pip3 install --no-cache-dir -e ./stac_fastapi/sfeos_helpers && \
     pip3 install --no-cache-dir ./stac_fastapi/opensearch[server]
 
 USER root
