+++ dockerfiles/Dockerfile.deploy.es
@@ -13,6 +13,7 @@ WORKDIR /app
 COPY . /app
 
 RUN pip install --no-cache-dir -e ./stac_fastapi/core
+RUN pip install --no-cache-dir -e ./stac_fastapi/sfeos_helpers
 RUN pip install --no-cache-dir ./stac_fastapi/elasticsearch[server]
 
 EXPOSE 8080
