+++ dockerfiles/Dockerfile.deploy.os
@@ -13,6 +13,7 @@ WORKDIR /app
 COPY . /app
 
 RUN pip install --no-cache-dir -e ./stac_fastapi/core
+RUN pip install --no-cache-dir -e ./stac_fastapi/sfeos_helpers
 RUN pip install --no-cache-dir ./stac_fastapi/opensearch[server]
 
 EXPOSE 8080
