+++ stac_fastapi/core/stac_fastapi/core/utilities.py
@@ -12,46 +12,6 @@ from stac_fastapi.types.stac import Item
 MAX_LIMIT = 10000
 
 
-def validate_refresh(value: Union[str, bool]) -> str:
-    """
-    Validate the `refresh` parameter value.
-
-    Args:
-        value (Union[str, bool]): The `refresh` parameter value, which can be a string or a boolean.
-
-    Returns:
-        str: The validated value of the `refresh` parameter, which can be "true", "false", or "wait_for".
-    """
-    logger = logging.getLogger(__name__)
-
-    # Handle boolean-like values using get_bool_env
-    if isinstance(value, bool) or value in {
-        "true",
-        "false",
-        "1",
-        "0",
-        "yes",
-        "no",
-        "y",
-        "n",
-    }:
-        is_true = get_bool_env("DATABASE_REFRESH", default=value)
-        return "true" if is_true else "false"
-
-    # Normalize to lowercase for case-insensitivity
-    value = value.lower()
-
-    # Handle "wait_for" explicitly
-    if value == "wait_for":
-        return "wait_for"
-
-    # Log a warning for invalid values and default to "false"
-    logger.warning(
-        f"Invalid value for `refresh`: '{value}'. Expected 'true', 'false', or 'wait_for'. Defaulting to 'false'."
-    )
-    return "false"
-
-
 def get_bool_env(name: str, default: Union[bool, str] = False) -> bool:
     """
     Retrieve a boolean value from an environment variable.
