+++ stac_fastapi/elasticsearch/setup.py
@@ -7,6 +7,7 @@ with open("README.md") as f:
 
 install_requires = [
     "stac-fastapi-core==4.1.0",
+    "sfeos-helpers==4.1.0",
     "elasticsearch[async]~=8.18.0",
     "uvicorn~=0.23.0",
     "starlette>=0.35.0,<0.36.0",
