+++ stac_fastapi/elasticsearch/stac_fastapi/elasticsearch/config.py
@@ -10,7 +10,7 @@ from elasticsearch._async.client import AsyncElasticsearch
 
 from elasticsearch import Elasticsearch  # type: ignore[attr-defined]
 from stac_fastapi.core.base_settings import ApiBaseSettings
-from stac_fastapi.core.utilities import get_bool_env, validate_refresh
+from stac_fastapi.sfeos_helpers.utilities import get_bool_env, validate_refresh
 from stac_fastapi.types.config import ApiSettings
