+++ stac_fastapi/sfeos_helpers/setup.cfg
@@ -1,0 +1,2 @@
+[metadata]
+version = attr: stac_fastapi.sfeos_helpers.version.__version__
