+++ stac_fastapi/sfeos_helpers/setup.py
@@ -1,4 +1,4 @@
-"""stac_fastapi: opensearch module."""
+"""stac_fastapi: helpers elasticsearch/ opensearch module."""
 
 from setuptools import find_namespace_packages, setup
 
@@ -6,31 +6,12 @@ with open("README.md") as f:
     desc = f.read()
 
 install_requires = [
-    "stac-fastapi-core==4.1.0",
-    "sfeos-helpers==4.1.0",
-    "opensearch-py~=2.8.0",
-    "opensearch-py[async]~=2.8.0",
-    "uvicorn~=0.23.0",
-    "starlette>=0.35.0,<0.36.0",
+    "stac-fastapi.core==4.1.0",
 ]
 
-extra_reqs = {
-    "dev": [
-        "pytest~=7.0.0",
-        "pytest-cov~=4.0.0",
-        "pytest-asyncio~=0.21.0",
-        "pre-commit~=3.0.0",
-        "requests>=2.32.0,<3.0.0",
-        "ciso8601~=2.3.0",
-        "httpx>=0.24.0,<0.28.0",
-    ],
-    "docs": ["mkdocs~=1.4.0", "mkdocs-material~=9.0.0", "pdocs~=1.2.0"],
-    "server": ["uvicorn[standard]~=0.23.0"],
-}
-
 setup(
-    name="stac_fastapi_opensearch",
-    description="Opensearch stac-fastapi backend.",
+    name="sfeos_helpers",
+    description="Helper library for the Elasticsearch and Opensearch stac-fastapi backends.",
     long_description=desc,
     long_description_content_type="text/markdown",
     python_requires=">=3.9",
@@ -50,8 +31,4 @@ setup(
     packages=find_namespace_packages(),
     zip_safe=False,
     install_requires=install_requires,
-    extras_require=extra_reqs,
-    entry_points={
-        "console_scripts": ["stac-fastapi-opensearch=stac_fastapi.opensearch.app:run"]
-    },
 )
