+++ stac_fastapi/sfeos_helpers/stac_fastapi/sfeos_helpers/version.py
@@ -1,0 +1,2 @@
+"""library version."""
+__version__ = "4.1.0"
